LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/history.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/go-back.html

Harness Error (TIMEOUT), message = null

TIMEOUT back to same-origin-initiated navigation Test timed out
TIMEOUT back to same-site-initiated navigation Test timed out
TIMEOUT back to cross-site-initiated navigation Test timed out
FAIL back to same-origin-initiated navigation 1: sec-fetch-dest assert_equals: expected "document" but got ""
FAIL back to same-origin-initiated navigation 1: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL back to same-origin-initiated navigation 1: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS back to same-origin-initiated navigation 1: sec-fetch-user
